+++ ui/mods/com.pa.quitch.gwaioverhaul/units/icon.js
@@ -1,0 +1,1 @@
+model.strategicIcons(model.strategicIcons().concat(["bot_aa"]));