+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - General Commander's minions are now dealt as cards to your inventory
 - Added Bot Anti-Air Tech card
+- Fixed armies not always landing with multiple commanders even though intelligence said they have them
 
 ## v4.2.2 - 2020-08-07